requirements.txt
Django==1.7.2
django-crispy-forms==1.6.1
